requirements.txt
# ========================================
# LiberALL Bot - Stack Oficial
# Python 3.11 + aiogram 3.3.0 + Firebase Firestore
# ========================================

# Core Dependencies - Telegram Bot
aiogram==3.3.0

# Firebase - Banco de Dados e Storage
firebase-admin==6.4.0

# Security & Encryption
cryptography==41.0.7
bcrypt==4.1.2
PyJWT==2.8.0

# Image Processing
Pillow==10.1.0
opencv-python==4.8.1.78
exifread==3.0.0
cloudinary==1.36.0
moviepy==1.0.3

# HTTP & Async Support
requests==2.31.0
aiofiles==23.2.1
aiohttp==3.9.1

# Environment & Configuration
python-dotenv==1.0.0

# Data Processing
pandas==2.1.4
numpy==1.25.2

# Logging and Monitoring
structlog==23.2.0
psutil==5.9.6

# Payment Processing (Optional)
stripe==7.8.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
validators==0.22.0

# Development Dependencies (Optional)
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1